+++ lib/vagrant/downloaders/http.rb
@@ -15,8 +15,10 @@ module Vagrant
       end
 
       def download!(source_url, destination_file)
+        proxy_uri = URI.parse(ENV["http_proxy"] || "")
         uri = URI.parse(source_url)
-        http = Net::HTTP.new(uri.host, uri.port)
+        http = Net::HTTP.new(uri.host, uri.port, proxy_uri.host, proxy_uri.port, proxy_uri.user, proxy_uri.password)
+
         if uri.scheme == "https"
           http.use_ssl = true
           http.verify_mode = OpenSSL::SSL::VERIFY_NONE